fixed merging conflicts
