+++ client/components/navbar.js
@@ -20,32 +20,30 @@ const Navbar = ({handleClick, isLoggedIn}) => (
       {isLoggedIn ? (
         <div>
           {/* The navbar will show these links after you log in */}
-          <p>
-            <Link to="/">
-              <img
-                src="https://cdn5.vectorstock.com/i/1000x1000/60/34/whiskey-barrel-premium-quality-wheat-wreath-flat-vector-13806034.jpg"
-                width={50}
-              />
-            </Link>
-            <Link to="/catalog">Catalog</Link>
-            <Link to="/checkout">
-              <img
-                src="https://pngriver.com/wp-content/uploads/2018/04/Download-Shopping-Cart-Logo-Png-Image-78016-For-Designing-Projects.png"
-                width={50}
-              />
-            </Link>
-            <a href="#" onClick={handleClick}>
-              Logout
-            </a>
-          </p>
+          <Link to="/">
+            <img
+              src="https://www.uncommongoods.com/images/items/23900/23960_1_640px.jpg"
+              width={50}
+            />
+          </Link>
+          <Link to="/catalog">Catalog</Link>
+          <Link to="/checkout">
+            <img
+              src="https://pngriver.com/wp-content/uploads/2018/04/Download-Shopping-Cart-Logo-Png-Image-78016-For-Designing-Projects.png"
+              width={50}
+            />
+          </Link>
+          <a href="#" onClick={handleClick}>
+            Logout
+          </a>
         </div>
       ) : (
         <div id="navBar">
           {/* The navbar will show these links before you log in */}
           <Link to="/">
             <img
-              src="https://cdn5.vectorstock.com/i/1000x1000/60/34/whiskey-barrel-premium-quality-wheat-wreath-flat-vector-13806034.jpg"
-              width={50}
+              src="https://www.uncommongoods.com/images/items/23900/23960_1_640px.jpg"
+              width={75}
             />
           </Link>
           <Link to="/catalog">Catalog</Link>
