+++ client/store/index.js
@@ -4,8 +4,9 @@ import thunkMiddleware from 'redux-thunk';
 import {composeWithDevTools} from 'redux-devtools-extension';
 import user from './user';
 import item from './item';
+import order from './order';
 
-const reducer = combineReducers({user, item});
+const reducer = combineReducers({user, item, order});
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
 );
@@ -14,3 +15,4 @@ const store = createStore(reducer, middleware);
 export default store;
 export * from './user';
 export * from './item';
+export * from './order';
