+++ client/store/item.js
@@ -5,21 +5,17 @@ import axios from 'axios';
  */
 const GET_ITEMS = 'GET_ITEMS';
 const GET_SELECTED_ITEM = 'GET_SELECTED_ITEM';
-const IN_CART = 'IN_CART';
-const ADD_TO_CART = 'ADD_TO_CART';
 
 /**
  * INITIAL STATE
  */
-const defaultItemList = {items: [], item: {}, cart: []};
+const defaultItemList = {items: [], item: {}};
 
 /**
  * ACTION CREATORS
  */
 const getItems = items => ({type: GET_ITEMS, items});
 const getItem = item => ({type: GET_SELECTED_ITEM, item});
-const inCart = cart => ({type: IN_CART, cart});
-const addingToCart = item => ({type: ADD_TO_CART, item});
 
 /**
  * THUNK CREATORS
@@ -43,27 +39,6 @@ export const getSelectedItem = id => async dispatch => {
   }
 };
 
-export const insideCart = userId => async dispatch => {
-  try {
-    const {data} = await axios.get(`/api/orders/${userId}/cart`);
-    dispatch(inCart(data));
-  } catch (error) {
-    console.error(error);
-  }
-};
-
-export const addToCart = (itemId, quanity, userId) => async dispatch => {
-  try {
-    const {data} = await axios.post(`/api/orders/${userId}/addcart`, {
-      itemId,
-      quanity
-    });
-    dispatch(addingToCart(data));
-  } catch (error) {
-    console.error(error);
-  }
-};
-
 /**
  * REDUCER
  */
@@ -73,10 +48,6 @@ export default function(state = defaultItemList, action) {
       return {...state, items: action.items};
     case GET_SELECTED_ITEM:
       return {...state, item: action.item};
-    case IN_CART:
-      return {...state, cart: action.cart};
-    case ADD_TO_CART:
-      return {...state, cart: [...state.cart, action.item]};
     default:
       return state;
   }
